configure monaco editor
