+++ packages/blocks-ui/src/canvas.js
@@ -4,8 +4,6 @@ import { jsx } from 'theme-ui'
 import prettier from 'prettier/standalone'
 import parserJS from 'prettier/parser-babylon'
 
-import { Clipboard, Check } from 'react-feather'
-
 import CodeEditor from './code-editor'
 
 import * as transforms from './transforms'
@@ -13,8 +11,6 @@ import { useEditor } from './providers/editor'
 import { useCode } from './providers/code'
 import InlineRender from './inline-render'
 import { PreviewArea, Device } from './device-preview'
-import { IconButton } from './ui'
-import useCopyToClipboard from './use-copy-to-clipboard'
 import { useScope } from './providers/scope'
 import { useCanvas } from './providers/canvas'
 import { useElementSize } from './use-element-size'
@@ -41,47 +37,36 @@ const CanvasWrap = props => {
   )
 }
 
-const Copy = ({ toCopy }) => {
-  const { hasCopied, copyToClipboard } = useCopyToClipboard()
-
-  return (
-    <IconButton
-      onClick={() => copyToClipboard(toCopy)}
-      sx={{ position: 'absolute', right: '-4px' }}
-    >
-      {hasCopied ? (
-        <Check sx={{ color: 'green' }} aria-label="Copied" />
-      ) : (
-        <Clipboard size={16} aria-label="Copy" />
-      )}
-    </IconButton>
-  )
-}
-
 const Canvas = () => {
-  const [error, setError] = useState(null)
   const { theme, ...scope } = useScope()
   const { code, transformedCode, editCode } = useCode()
   const { mode } = useEditor()
-  const rawCode = transforms.toRawJSX(code)
-  const formattedCode = prettier.format(rawCode, {
-    parser: 'babel',
-    plugins: [parserJS]
-  })
+
+  const [formattedCode, setFormattedCode] = useState('')
+
+  useEffect(() => {
+    if (mode === 'code') {
+      const rawCode = transforms.toRawJSX(code)
+      setFormattedCode(
+        prettier.format(rawCode, {
+          parser: 'babel',
+          plugins: [parserJS]
+        })
+      )
+    }
+  }, [mode])
 
   const onCodeChange = code => {
     try {
       editCode(code)
-      setError(null)
     } catch (e) {
-      setError(e.toString())
+      // do nothing as they are displayed in the editor
     }
   }
 
   if (mode === 'code') {
     return (
       <CanvasWrap>
-        <Copy toCopy={formattedCode} />
         <CodeEditor code={formattedCode} onChange={onCodeChange} />
       </CanvasWrap>
     )
